requirements.txt
termcolor==1.1.0
ciso8601==2.2.0
python-dotenv==0.19.2
Telethon==1.24.0
ccxt==1.77.80
pandas==1.4.2
pytz==2022.1
numpy==1.22.3 